feat: image upload added for map object

+++ src/Modules/Tours/Explorer.Tours.Core/UseCases/Administration/MapObjectService.cs
@@ -30,10 +30,17 @@ namespace Explorer.Tours.Core.UseCases.Administration
 
         public Result DeleteObjectAndRequest(int mapObjectId)
         {
-            var request = _internalObjectRequestService.GetRequestByMapObjectId(mapObjectId);
-            if (request == null) throw new Exception($"Request for MapObject with ID {mapObjectId} not found.");
-            _internalObjectRequestService.Delete(request.Value.Id);
-            return Delete(mapObjectId);
+            try
+            {
+                var request = _internalObjectRequestService.GetRequestByMapObjectId(mapObjectId);
+                if (request == null) throw new Exception($"Request for MapObject with ID {mapObjectId} not found.");
+                if (request.Value != null) _internalObjectRequestService.Delete(request.Value.Id);
+                return Delete(mapObjectId);
+            }
+            catch (KeyNotFoundException e)
+            {
+                return Result.Fail(FailureCode.NotFound).WithError(e.Message);
+            }
         }
     }
 }
